# contracts/atracciones-app-openapi_contrato.yaml
openapi: 3.0.3
info:
  title: API de Atracciones Turísticas - Contrato de la aplicación implementada
  version: 2.4.0
  description: |
    Contrato de la aplicación **Atracciones Web** tal como está implementada y desplegada.

    Incluye las 14 operaciones del contrato del grupo (`atracciones-openapi.yaml`, v1.2.0)
    más los endpoints propios de la aplicación: autenticación (`/auth/*`), reportes del operador (`/reportes/*`)
    y gestión de administradores (`/administradores`, solo para el administrador principal).

    Extensión a una operación del contrato: `POST /atracciones/search` acepta el campo opcional `query` (texto libre).
    Es compatible hacia atrás: quien no lo envía obtiene el mismo resultado que con el contrato del grupo.

    Regla de horarios (aplicada por la aplicación): para la fecha de hoy solo se pueden comprar los horarios que empiezan
    más de 30 minutos después de la hora actual. La hora es la local de la atracción (Ecuador continental UTC-5,
    Galápagos UTC-6). Los horarios están en formato de 24 horas (`HH:MM`).

    **Autenticación.** La API emite su propio token JWT (Bearer). Se obtiene en `POST /auth/login`,
    `POST /auth/registro` o `POST /auth/google`. Los permisos (scopes) usan los mismos nombres del contrato del grupo.

    | Rol | Scopes |
    |---|---|
    | CLIENTE | attractions:read, attractions:book, attractions:cancel |
    | OPERADOR | attractions:read, attractions:cancel, attractions:write (no puede reservar) |

...

...

...

...

...

...
servers:
...
paths:
...
  /atracciones/search:
    post:
      tags: [Atracciones - Catálogo]
      summary: Búsqueda de atracciones
      description: |
        Búsqueda con filtros por ciudad, país, fechas, calificación, orden y paginación por token.

        Extensión de la aplicación: el campo opcional `query` busca texto libre en el nombre, la descripción,
        las categorías y la dirección. No distingue mayúsculas ni tildes (`galápagos` = `GALAPAGOS`).
        Si trae varias palabras (máximo 5), todas deben aparecer. Se combina con el resto de los filtros.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAtraccionesRequest'
      responses:
        '200':
          description: Resultados de la búsqueda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAtraccionesResponse'
        '400':
          $ref: '#/components/responses/Problem400'
  /atracciones/details:
    post:
      tags: [Atracciones - Catálogo]
      summary: Detalles de múltiples atracciones (por lote)
      description: Recibe hasta 50 identificadores y evita consultas una por una.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetailsRequest'
      responses:
        '200':
          description: Detalles de las atracciones pedidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAtraccionesResponse'
        '400':
          $ref: '#/components/responses/Problem400'
  /atracciones:
    get:
      tags: [Atracciones - Catálogo]
      summary: Listado paginado de atracciones
      description: |
        Ordenado por nombre ascendente. Paginación con `limit` y `offset`.
        Extensión de la aplicación: el parámetro opcional `q` filtra por texto (lo usa el buscador del catálogo del panel del operador).
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: q
          in: query
          required: false
          description: |
            Extensión de la aplicación (no está en el contrato del grupo). Texto libre que se busca en nombre,
            descripción, categorías y dirección, sin distinguir mayúsculas ni tildes. Con varias palabras,
            todas deben aparecer. Vacío o solo espacios no filtra nada.
          schema:
            type: string
            maxLength: 100
          example: frailes
      responses:
        '200':
          description: Listado recuperado
          headers:
            X-API-Deprecation-Date:
              schema:
                type: string
              description: Fecha en que esta versión será deprecada.
            Cache-Control:
              schema:
                type: string
              description: Directivas de caché (max-age=300).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAtraccionResponse'
        '400':
          $ref: '#/components/responses/Problem400'
    post:
      tags: [Atracciones - Catálogo]
      summary: Registrar una nueva atracción
      security:
        - bearerAuth: [attractions:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAtraccionRequest'
      responses:
        '201':
          description: Atracción creada
          headers:
            Location:
              schema:
                type: string
              description: Ruta de la atracción creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtraccionResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
  /atracciones/{id}:
    parameters:
      - $ref: '#/components/parameters/AtraccionId'
    get:
      tags: [Atracciones - Catálogo]
      summary: Detalle de una atracción
      security: []
      responses:
        '200':
          description: Detalle de la atracción
          headers:
            X-API-Deprecation-Date:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtraccionResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '404':
          $ref: '#/components/responses/Problem404'
    put:
      tags: [Atracciones - Catálogo]
      summary: Reemplazar los datos completos de una atracción
      security:
        - bearerAuth: [attractions:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAtraccionRequest'
      responses:
        '204':
          description: Reemplazo exitoso, sin contenido
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
        '404':
          $ref: '#/components/responses/Problem404'
    patch:
      tags: [Atracciones - Catálogo]
      summary: Actualizar parcialmente una atracción
      security:
        - bearerAuth: [attractions:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAtraccionRequest'
      responses:
        '200':
          description: Actualización exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtraccionResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
        '404':
          $ref: '#/components/responses/Problem404'
    delete:
      tags: [Atracciones - Catálogo]
      summary: Eliminar una atracción
      description: Borrado lógico (soft delete). La atracción deja de aparecer, pero sus reservas anteriores se conservan.
      security:
        - bearerAuth: [attractions:write]
      responses:
        '204':
          description: Eliminación exitosa, sin contenido
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
        '404':
          $ref: '#/components/responses/Problem404'
  /atracciones/{id}/availability:
    get:
      tags: [Atracciones - Catálogo]
      summary: Consultar la disponibilidad de cupos
      description: |
        Cupos disponibles = cupo diario de la atracción menos los tickets de reservas no canceladas en esa fecha.

        `times` solo trae los horarios que todavía se pueden comprar: para hoy se quitan los que ya empezaron o empiezan
        en menos de 30 minutos (hora local de la atracción). Si la atracción tiene horarios y ya cerraron todos, o la fecha
        ya pasó, `available_spots` es 0.
      security: []
      parameters:
        - $ref: '#/components/parameters/AtraccionId'
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: '2026-10-10'
      responses:
        '200':
          description: Disponibilidad recuperada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '404':
          $ref: '#/components/responses/Problem404'
  /atracciones/{id}/reservations:
    post:
      tags: [Atracciones - Reservas]
      summary: Crear una reserva
      description: |
        Solo el rol CLIENTE puede reservar. Se ejecuta en una transacción con bloqueo de la atracción, por lo que dos compras
        simultáneas no pueden vender el mismo último cupo. Valida que la fecha no haya pasado, que el horario exista, que
        el horario de hoy no haya cerrado su venta (30 minutos antes de empezar, hora local de la atracción) y que haya cupos.
        Si la atracción tiene horarios y la reserva es para hoy, `time` es obligatorio. La reserva queda en estado
        CONFIRMED y el total se calcula en USD.
      security:
        - bearerAuth: [attractions:book]
      parameters:
        - $ref: '#/components/parameters/AtraccionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '201':
          description: Reserva confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
        '404':
          $ref: '#/components/responses/Problem404'
        '409':
          $ref: '#/components/responses/Problem409'
  /atracciones/reservations:
    get:
      tags: [Atracciones - Reservas]
      summary: Historial de reservas
      description: El cliente ve solo sus reservas; el operador ve todas. Ordenado de la más reciente a la más antigua.
      security:
        - bearerAuth: [attractions:read]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Listado de reservas
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total de reservas disponibles (para paginar).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
  /atracciones/reservations/{reservationId}:
    get:
      tags: [Atracciones - Reservas]
      summary: Detalle de una reserva
      description: Un cliente que pide una reserva ajena recibe 404, igual que si no existiera.
      security:
        - bearerAuth: [attractions:read]
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      responses:
        '200':
          description: Detalle de la reserva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
        '404':
          $ref: '#/components/responses/Problem404'
  /atracciones/reservations/{reservationId}/cancel:
    post:
      tags: [Atracciones - Reservas]
      summary: Cancelar una reserva
      description: |
        El cliente solo puede cancelar sus propias reservas; el operador puede cancelar cualquiera.
        Devuelve 409 si la reserva ya estaba cancelada o si la atracción no tiene cancelación gratuita.
      security:
        - bearerAuth: [attractions:cancel]
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelReservationRequest'
      responses:
        '200':
          description: Cancelación procesada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
        '404':
          $ref: '#/components/responses/Problem404'
        '409':
          $ref: '#/components/responses/Problem409'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT emitido por esta misma API (emisor `atracciones-api`).
        Contenido: sub (id), email, name, role y scope. Vigencia por defecto: 3600 segundos.
        Scopes: attractions:read, attractions:book, attractions:cancel, attractions:write.
  parameters:
    AtraccionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReservationId:
      name: reservationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID generado por el cliente para esta operación. Si falta o no es un UUID, responde 400.
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      description: Elementos por página
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Offset:
      name: offset
      in: query
      description: Elementos a saltar desde el inicio
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Problem400:
      description: Petición inválida (datos, formato o cabecera Idempotency-Key incorrectos)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Problem401:
      description: Falta el token, es inválido o está vencido (o las credenciales son incorrectas)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Problem403:
      description: El token es válido pero no tiene el permiso (scope) necesario
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Problem404:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Problem409:
      description: Conflicto (sin cupos, reserva ya cancelada, sin cancelación gratuita, correo ya registrado o error de idempotencia)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    Operator:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: Quito Tour Bus
...
    DetailsRequest:
      type: object
      required: [attractions]
      properties:
        attractions:
          type: array
          maxItems: 50
          items:
            type: string
        languages:
          type: array
          items:
            type: string
    SearchAtraccionesRequest:
      type: object
      properties:
        currency:
          type: string
          example: USD
          description: Solo se acepta USD
        cities:
          type: array
          items:
            type: integer
        countries:
          type: array
          items:
            type: string
          example: [ec]
        query:
          type: string
          maxLength: 100
          example: playa
          description: |
            Extensión de la aplicación (no está en el contrato del grupo). Texto libre que se busca en nombre,
            descripción, categorías y dirección, sin distinguir mayúsculas ni tildes. Con varias palabras,
            todas deben aparecer. Un texto vacío o solo con espacios no filtra nada.
        dates:
          type: object
          required: [start_date, end_date]
          properties:
            start_date:
              type: string
              format: date
            end_date:
              type: string
              format: date
        filters:
          type: object
          properties:
            rating:
              type: object
              properties:
                minimum_review_score:
                  type: number
                minimum_review_count:
                  type: integer
        next_page:
          type: string
          description: Token opaco de paginación
        rows:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        sort:
          type: object
          required: [by]
          properties:
            by:
              type: string
              enum: [most_popular, top_rated, price_asc, price_desc]
    SearchAtraccionesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AtraccionResponse'
        metadata:
          type: object
          properties:
            total_results:
              type: integer
            next_page:
              type: string
        request_id:
          type: string
    CreateAtraccionRequest:
      type: object
      required: [name, long_description, duration, price, operator, product_type, includes, categories, locations, photos, supported_languages, free_cancellation]
      properties:
        name:
          type: string
          minLength: 3
        long_description:
          type: string
          minLength: 10
        duration:
          type: string
          description: Duración en formato ISO 8601
          example: PT8H
        price:
          $ref: '#/components/schemas/Price'
        operator:
          $ref: '#/components/schemas/Operator'
        product_type:
          $ref: '#/components/schemas/ProductType'
        includes:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
        badges:
          type: array
          items:
            type: string
        locations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Location'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        supported_languages:
          type: array
          items:
            type: string
        free_cancellation:
          type: boolean
    UpdateAtraccionRequest:
      type: object
      description: Mismos campos que CreateAtraccionRequest, todos opcionales.
      properties:
        name:
          type: string
          minLength: 3
        long_description:
          type: string
          minLength: 10
        duration:
          type: string
        price:
          $ref: '#/components/schemas/Price'
        operator:
          $ref: '#/components/schemas/Operator'
        product_type:
          $ref: '#/components/schemas/ProductType'
        includes:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
        badges:
          type: array
          items:
            type: string
        locations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Location'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        supported_languages:
          type: array
          items:
            type: string
        free_cancellation:
          type: boolean
    AtraccionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        long_description:
          type: string
        duration:
          type: string
        price:
          $ref: '#/components/schemas/Price'
        operator:
          $ref: '#/components/schemas/Operator'
        product_type:
          $ref: '#/components/schemas/ProductType'
        includes:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
        badges:
          type: array
          items:
            type: string
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        supported_languages:
          type: array
          items:
            type: string
        free_cancellation:
          type: boolean
        ratings:
          $ref: '#/components/schemas/Rating'
        url:
          $ref: '#/components/schemas/Url'
        _links:
          type: object
          description: Enlaces HATEOAS (self, actualizar, eliminar)
          additionalProperties:
            type: object
            properties:
              href:
                type: string
              method:
                type: string
    PaginatedAtraccionResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AtraccionResponse'
        meta:
          type: object
          properties:
            totalItems:
              type: integer
            itemCount:
              type: integer
            itemsPerPage:
              type: integer
            totalPages:
              type: integer
            currentPage:
              type: integer
        _links:
          type: object
          properties:
            first:
              type: string
            previous:
              type: string
            next:
              type: string
            last:
              type: string
    AvailabilityResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        available_spots:
          type: integer
          minimum: 0
        times:
          type: array
          items:
            type: string
          example: ['10:00', '14:00']
          description: Horarios en formato de 24 horas que todavía se pueden comprar en la fecha consultada.
    ReservationRequest:
      type: object
      required: [date, ticket_count, customer_name]
      properties:
        date:
          type: string
          format: date
          description: No puede ser una fecha pasada (se compara con la fecha local de la atracción).
        time:
          type: string
          example: '10:00'
          description: |
            Debe ser uno de los horarios de la atracción, en formato de 24 horas (`HH:MM`). Para la fecha de hoy debe empezar
            más de 30 minutos después de la hora local de la atracción.
        ticket_count:
          type: integer
          minimum: 1
        customer_name:
          type: string
        customer_email:
          type: string
          format: email
    ReservationResponse:
      type: object
      properties:
        reservation_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [CONFIRMED, PENDING, CANCELLED]
        ticket_count:
          type: integer
        total_price:
          $ref: '#/components/schemas/Price'
    CancelReservationRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          example: Plan cancelado
...